+++ example/lib/feature_b/route.dart
@@ -8,7 +8,7 @@ class FeatureBRoute extends NavigationRouteHandler {
   bool canOpen(String path) => path == '/feature-b';
 
   @override
-  Route<T> buildRoute<T>(String path, [RouteArguments? arguments]) {
+  Route<void> buildRoute(String path, [RouteArguments? arguments]) {
     return MaterialPageRoute(
       builder: (context) => FeatureBHomePage(
         openedBy: arguments?['opened_by'],