+++ lib/src/errors.dart
@@ -3,8 +3,8 @@ import 'route_arguments.dart';
 /// Exception thrown when attempting to open a route but none registered were
 /// capable of opening it.
 ///
-/// See [Router] and [RouteHandler] for more information about how to go about
-/// registering and opening routes.
+/// See [PathRouter] and [RouteHandler] for more information about how to go
+/// about registering and opening routes.
 class UnregisteredRouteException implements Exception {
   const UnregisteredRouteException(this.route, [this.arguments]) : super();